Set Deno extension as the JSON formatter
The built-in VS Code JSON Language Features extension `vscode.json-language-features`
cannot fix missing commas or quotation marks in JSON. It also seems to have a strict
parser which outputs a messy, unformatted JSON file whenever it encounters an error.

The Prettier extension `esbenp.prettier-vscode` can somewhat format JSON, but to the
best of my knowledge and experimentation, it cannot fix missing commas or quotation marks.

The Deno extension `denoland.vscode-deno`, however, does fix missing commas and quotation marks in JSON.

If setting `editor.defaultFormatter: "denoland.vscode-deno"` in `settings.json`
does not seem to work, a workaround is to open a JSON file, open the
command palette to run "Format Document With...", and select "Deno", which adds the entry to your `settings.json` for you.